norme

+++ srcs/game/ft_print_map.c
@@ -6,26 +6,28 @@
 /*   By: ajimenez <ajimenez@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/15 10:29:16 by ajimenez          #+#    #+#             */
-/*   Updated: 2021/12/15 18:59:30 by ajimenez         ###   ########.fr       */
+/*   Updated: 2021/12/16 18:02:11 by ajimenez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/so_long.h"
 
-static void	ft_print_coin(t_vars *mlx, size_t x, size_t y)
+static void	ft_print_coin(t_vars *mlx, int x, int y)
 {
 	mlx_put_image_to_window(mlx->ptr, mlx->win, mlx->coin.ptr, 64 * x, 64 * y);
 	mlx->coin.count++;
 }
 
-static void	ft_print_mario(t_vars *mlx, size_t x, size_t y)
+static void	ft_print_mario(t_vars *mlx, int x, int y)
 {
 	mlx_put_image_to_window(mlx->ptr, mlx->win, mlx->mario.ptr, 64 * x, 64 * y);
 	mlx->mario.x = x;
 	mlx->mario.y = y;
+//	printf("y ==> %d\n", mlx->mario.y);
+//	printf("x ==> %d\n", mlx->mario.x);
 }
 
-static	void	ft_print_map(char **map, t_vars *mlx, size_t x, size_t y)
+static	void	ft_print_map(char **map, t_vars *mlx, int x, int y)
 {
 	if (map[y][x] == '1')
 		mlx_put_image_to_window(mlx->ptr, mlx->win, mlx->block, 64 * x, 64 * y);
@@ -44,14 +46,14 @@ static	void	ft_print_map(char **map, t_vars *mlx, size_t x, size_t y)
 
 void	ft_iter_map(t_map *map, t_vars *mlx)
 {
-	size_t	x;
-	size_t	y;
+	int	x;
+	int	y;
 
 	y = 0;
-	while (y < (size_t )map->data.line_count)
+	while (y < map->data.line_count)
 	{
 		x = 0;
-		while(x < (size_t )map->data.max)
+		while(x < map->data.max)
 		{
 			ft_print_map(map->map, mlx, x, y);
 			x++;
